requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
pandas>=1.5
numpy>=1.24
sqlalchemy>=2.0
mysql-connector-python>=8.0
python-dotenv
openai
pydantic
tqdm
